Add comments and generate CZICompress_Config.h
Updated `mega-linter.yml` to include comments explaining the generation of the `CZICompress_Config.h` file. This file is essential for the linters to operate correctly, as it is usually created during the CMake build process but is absent from the repository. A new step has been added to create this file, preventing linting issues.